refactor: extract WantedScanner class to services/wanted_scanner_core.py
- Move full WantedScanner class (~1085 LOC) to wanted_scanner_core.py
- wanted_scanner.py becomes thin facade (~86 LOC) re-exporting public API
- FULL_SCAN_INTERVAL constant kept in core and re-exported from facade
- All 10 external callers unaffected (still import from services.wanted_scanner)
- Add smoke tests confirming public API survives the split
